refactor: keep D1 migrations in SQL
