+++ package.json
@@ -15,9 +15,6 @@
     "build": "vite build && node scripts/build-pwa.mjs && node scripts/check-setup-preview-boundary.mjs",
     "deploy": "node scripts/release/deploy.mjs",
     "db:reset:local": "wrangler d1 execute DB --local --yes --file scripts/hqbase/reset-d1.sql && wrangler d1 migrations apply DB --local",
-    "db:generate": "drizzle-kit generate --config drizzle.config.ts",
-    "db:generate:custom": "drizzle-kit generate --config drizzle.config.ts --custom",
-    "db:check": "drizzle-kit check --config drizzle.config.ts && node scripts/check-drizzle-schema.mjs",
     "db:migrate:local": "wrangler d1 migrations apply DB --local",
     "db:migrate:remote": "wrangler d1 migrations apply DB --remote",
     "db:seed:local": "node scripts/seed-local.mjs",
@@ -38,7 +35,7 @@
     "test:e2e:staging:smoke": "playwright test --config playwright.config.ts app-shell-smoke.spec.ts",
     "test:e2e:staging:lifecycle": "playwright test --config playwright.config.ts lifecycle.spec.ts",
     "pwa:icons": "node scripts/generate-pwa-icons.mjs",
-    "check": "pnpm code:check && pnpm typecheck && pnpm api:check && pnpm db:check && pnpm test && pnpm test:coverage && pnpm test:architecture && pnpm build",
+    "check": "pnpm code:check && pnpm typecheck && pnpm api:check && pnpm test && pnpm test:coverage && pnpm test:architecture && pnpm build",
     "deploy:dry-run": "pnpm build && wrangler deploy --dry-run",
     "hqbase": "node scripts/hqbase/cli.mjs",
     "hqbase:button": "node scripts/hqbase/cli.mjs button",
@@ -115,7 +112,6 @@
     "@vitest/coverage-v8": "4.1.10",
     "autoprefixer": "^10.4.21",
     "cross-spawn": "7.0.6",
-    "drizzle-kit": "0.31.10",
     "postcss": "^8.5.23",
     "tailwindcss": "^3.4.17",
     "typescript": "^5.8.3",
