+++ scripts/hqbase/config.mjs
@@ -56,8 +56,7 @@ export function createWranglerConfig(manifest) {
         binding: "DB",
         database_name: manifest.d1.name,
         database_id: manifest.d1.id,
-        migrations_dir: `${rootFromDeployment}/migrations`,
-        migrations_pattern: `${rootFromDeployment}/migrations/**/*.sql`
+        migrations_dir: `${rootFromDeployment}/migrations`
       }
     ],
     r2_buckets: [
