+++ vitest.worker.config.ts
@@ -1,10 +1,9 @@
 import { fileURLToPath } from "node:url";
-import { cloudflareTest } from "@cloudflare/vitest-pool-workers";
+import { cloudflareTest, readD1Migrations } from "@cloudflare/vitest-pool-workers";
 import { defineConfig } from "vitest/config";
-import { readD1MigrationsRecursively } from "./scripts/d1-migrations.mjs";
 
 export default defineConfig(async () => {
-  const migrations = readD1MigrationsRecursively(
+  const migrations = await readD1Migrations(
     fileURLToPath(new URL("./migrations", import.meta.url))
   );
