+++ wrangler.jsonc
@@ -34,8 +34,7 @@
       "binding": "DB",
       "database_name": "hqbase",
       "database_id": "00000000-0000-0000-0000-000000000000",
-      "migrations_dir": "migrations",
-      "migrations_pattern": "migrations/**/*.sql"
+      "migrations_dir": "migrations"
     }
   ],
   "r2_buckets": [
